Several changes:
    - Handle iframe mouse-over events
    - Fix sizeHandle and fixedSize interaction in layout-manager
    - Delete resize-handles in grid-layout when needed
    - Unblocking all signals in child threads before exec (fixes #3222)
    - Handling FastCGI child process respawns outside of the signal handler
    (#3222)
    - Fixed ambiguous WLength overloads. Still allow for string literals.
    - Fixed include location for boost 1.56
    - Remove style-class info in widgetgallery example on button-click
    - Fixing up 3D chart examples, added subtitle class to blog
    - Fix WPaintDevice drawArc documentation
    - Fix case sensitivity issue in font matching (#3365)
    - Fixing the documentation of setPointSprite
    - PdfImage::drawArc cleanup (spanAngle here is already divided by 360)
    - Fix #3170 setDraggable and mouseWentUp are not compatible
    - Fix WCompositeWidget::setFirstFocus
    - Fix WPdfImage::drawArc correct radius and draw circle
    - Added run() shorthand (#2704)
    - Added WServer+setServerConfiguration shorthand constructor (#2702)
    - Going postal: added postAll (#3224)
    - Made the GL functions protected.
    - Do not restrict WSound to a single MP3 source
    - Removed some agentIsIE checks: point sprites work in IE 11 now
    - Removed hardcoded Float32Array
    - Changing the title font of chart axes had no effect
    - Chart3D: Fixed slightly transparent solid colored surfaces
    - Chart3D: fixed some allocation/deallocation code
    - Chart3D: antialiasing on/off switch support on webgl-experimental contexts.
    - Fixed a bug where the old session id was kept for file upload
    - added adjustSize() in WTableView
    - treeview-dragdrop example: restore numerical sorting

+++ src/Wt/WBoostAny.C
@@ -384,8 +384,8 @@ WString asString(const boost::any& v, const WT_USTRING& format)
   } else if (v.type() == typeid(boost::posix_time::time_duration)) {
     const boost::posix_time::time_duration& dt
       = boost::any_cast<boost::posix_time::time_duration>(v);
-    int millis = dt.fractional_seconds() *
-      std::pow(10.0, 3 - boost::posix_time::time_duration::num_fractional_digits()); 
+    int millis = (int)(dt.fractional_seconds() *
+      std::pow(10.0, 3 - boost::posix_time::time_duration::num_fractional_digits())); 
     return WTime(dt.hours(), dt.minutes(), dt.seconds(), millis)
       .toString(format.empty() ? "HH:mm:ss" : format);
   }
@@ -515,7 +515,7 @@ double asNumber(const boost::any& v)
   } else if (v.type() == typeid(boost::posix_time::time_duration)) {
     const boost::posix_time::time_duration& dt
       = boost::any_cast<boost::posix_time::time_duration>(v);
-    return dt.total_milliseconds();
+    return static_cast<double>(dt.total_milliseconds());
   }
 
 #define ELSE_NUMERICAL_ANY(TYPE) \
